+++ s2spy/__init__.py
@@ -5,9 +5,6 @@ This package is a high-level python package integrating expert knowledge
 and artificial intelligence to boost (sub) seasonal forecasting.
 """
 import logging
-from . import calendar_shifter
-from . import time
-from . import traintest
 from .rgdr.rgdr import RGDR
 
 
@@ -17,4 +14,4 @@ __author__ = "Yang Liu"
 __email__ = "y.liu@esciencecenter.nl"
 __version__ = "0.2.1"
 
-__all__ = ["time", "traintest", "RGDR", "calendar_shifter"]
+__all__ = ["RGDR"]